test: refactor: move `read_xor_key`/`util_xor` helpers to util module
